Changed the missed call control message logic
The control message is now clickable when the phone toggle is disabled or when the microphone permission is not given
