+++ app/src/main/java/org/thoughtcrime/securesms/database/MmsSmsColumns.java
@@ -41,7 +41,6 @@ public interface MmsSmsColumns {
     protected static final long MISSED_CALL_TYPE                   = 3;
     protected static final long JOINED_TYPE                        = 4;
     protected static final long FIRST_MISSED_CALL_TYPE             = 5;
-    protected static final long MISSED_PERMISSION_CALL_TYPE        = 6;
 
     protected static final long BASE_INBOX_TYPE                    = 20;
     protected static final long BASE_OUTBOX_TYPE                   = 21;
@@ -236,7 +235,7 @@ public interface MmsSmsColumns {
     public static boolean isCallLog(long type) {
       long baseType = type & BASE_TYPE_MASK;
       return baseType == INCOMING_CALL_TYPE || baseType == OUTGOING_CALL_TYPE ||
-              baseType == MISSED_CALL_TYPE || baseType == FIRST_MISSED_CALL_TYPE || baseType == MISSED_PERMISSION_CALL_TYPE;
+              baseType == MISSED_CALL_TYPE || baseType == FIRST_MISSED_CALL_TYPE;
     }
 
     public static boolean isExpirationTimerUpdate(long type) {
@@ -265,10 +264,6 @@ public interface MmsSmsColumns {
 
     public static boolean isMissedCall(long type) {
       return (type & BASE_TYPE_MASK) == MISSED_CALL_TYPE;
-    }
-
-    public static boolean isMissedPermissionCall(long type) {
-      return (type & BASE_TYPE_MASK) == MISSED_PERMISSION_CALL_TYPE;
     }
 
     public static boolean isFirstMissedCall(long type) {
