+++ app/src/main/java/org/thoughtcrime/securesms/database/model/DisplayRecord.java
@@ -142,9 +142,6 @@ public abstract class DisplayRecord {
   public boolean isFirstMissedCall() {
     return SmsDatabase.Types.isFirstMissedCall(type);
   }
-  public boolean isMissedPermissionCall() {
-    return SmsDatabase.Types.isMissedPermissionCall(type);
-  }
   public boolean isDeleted() { return  MmsSmsColumns.Types.isDeletedMessage(type); }
   public boolean isMessageRequestResponse() { return  MmsSmsColumns.Types.isMessageRequestResponse(type); }
