+++ app/src/main/java/org/thoughtcrime/securesms/database/model/MessageRecord.java
@@ -134,8 +134,6 @@ public abstract class MessageRecord extends DisplayRecord {
         callType = CallMessageType.CALL_OUTGOING;
       } else if (isMissedCall()) {
         callType = CallMessageType.CALL_MISSED;
-      } else if (isMissedPermissionCall()) {
-        callType = CallMessageType.CALL_MISSED_PERMISSION;
       } else {
         callType = CallMessageType.CALL_FIRST_MISSED;
       }
