+++ app/src/main/res/drawable/ic_incoming_call.xml
@@ -5,8 +5,8 @@
     android:viewportHeight="20">
   <path
       android:pathData="M14.414,7l3.293,-3.293a1,1 0,0 0,-1.414 -1.414L13,5.586V4a1,1 0,1 0,-2 0v4.003a0.996,0.996 0,0 0,0.617 0.921A0.997,0.997 0,0 0,12 9h4a1,1 0,1 0,0 -2h-1.586z"
-      android:fillColor="#000000"/>
+      android:fillColor="?message_received_text_color"/>
   <path
       android:pathData="M2,3a1,1 0,0 1,1 -1h2.153a1,1 0,0 1,0.986 0.836l0.74,4.435a1,1 0,0 1,-0.54 1.06l-1.548,0.773a11.037,11.037 0,0 0,6.105 6.105l0.774,-1.548a1,1 0,0 1,1.059 -0.54l4.435,0.74a1,1 0,0 1,0.836 0.986V17a1,1 0,0 1,-1 1h-2C7.82,18 2,12.18 2,5V3z"
-      android:fillColor="#000000"/>
+      android:fillColor="?message_received_text_color"/>
 </vector>
