+++ app/src/main/res/values/themes.xml
@@ -48,7 +48,7 @@
        <item name="menu_selectall_icon">@drawable/ic_baseline_select_all_24</item>
        <item name="menu_split_icon">@drawable/ic_baseline_call_split_24</item>
        <item name="menu_popup_expand">@drawable/ic_baseline_launch_24</item>
-       <item name="menu_info_icon">@drawable/ic_baseline_info_24</item>
+       <item name="menu_info_icon">@drawable/ic_info_outline_white_24dp</item>
        <item name="menu_pin_icon">@drawable/ic_outline_pin_24</item>
        <item name="menu_unpin_icon">@drawable/ic_outline_pin_off_24</item>
        <item name="menu_mark_all_as_read">@drawable/ic_outline_mark_chat_read_24</item>
