+++ libsession/src/main/java/org/session/libsession/messaging/calls/CallMessageType.kt
@@ -4,6 +4,5 @@ enum class CallMessageType {
     CALL_MISSED,
     CALL_INCOMING,
     CALL_OUTGOING,
-    CALL_FIRST_MISSED,
-    CALL_MISSED_PERMISSION,
+    CALL_FIRST_MISSED
 }
